+++ unstructured/partition/epub.py
@@ -13,6 +13,7 @@ def partition_epub(
     include_page_breaks: bool = False,
     include_metadata: bool = True,
     metadata_filename: Optional[str] = None,
+    metadata_date: Optional[str] = None,
     **kwargs,
 ) -> List[Element]:
     """Partitions an EPUB document. The document is first converted to HTML and then
@@ -25,7 +26,10 @@ def partition_epub(
     file
         A file-like object using "rb" mode --> open(filename, "rb").
     include_page_breaks
-        If True, the output will include page breaks if the filetype supports it.
+        If True, the output will include page breaks if the filetype supports it
+    metadata_date
+        The last modified date for the document.
+
     """
     return convert_and_partition_html(
         source_format="epub",
@@ -33,4 +37,5 @@ def partition_epub(
         file=file,
         include_page_breaks=include_page_breaks,
         metadata_filename=metadata_filename,
+        metadata_date=metadata_date,
     )