+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ### Enhancements
 
 * Add parameter `skip_infer_table_types` to enable (skip) table extraction for other doc types
+* Adds optional Unstructured API unit tests in CI
+* Tracks last modified date for all document types.
 
 ### Features
 
 ### Fixes
-
 
 ## 0.8.4